Allow the 'LDAP Admin Filter' to be cleared
Currently if you clear the LDAP Admin Filter field, it resets
to the default '(objectClass=JellyfinAdministrator)' which is
a little confusing.

This patch changes the code to set the state to '_disabled_' when
persisting to the plugin configuration. The backend then checks
for that value and only applies the LDAP filter if the field
does not equal '_disabled_'.

Signed-off-by: James Ketrenos <james_github@ketrenos.com>